+++ pkg/protocol/xprotocol/wasm/protocol.go
@@ -21,7 +21,7 @@ import (
 	"context"
 	"sync/atomic"
 
-	"mosn.io/mosn/pkg/wasm/abi/xproxywasm020"
+	"mosn.io/mosn/pkg/wasm/abi/ext/xproxywasm020"
 
 	"mosn.io/api"
 	"mosn.io/mosn/pkg/log"
